requirements.txt
fastapi
uvicorn
python-multipart
lightrag-hku
google-generativeai
numpy